chatbot/requirement text.txt
openai
PyPDF2
python-docx
bcrypt
scikit-learn
tk

